requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
bcrypt==4.1.2
psycopg2-binary==2.9.9
qrcode[pil]==7.4.2
Pillow==10.1.0
openpyxl==3.1.2
pandas==2.1.4
python-dotenv==1.0.0
twilio==8.10.0
beautifulsoup4==4.12.2
requests==2.31.0
weasyprint==60.1
